requirement.txt
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
psycopg2==2.9.9
sqlparse==0.5.0
